Remove wargame from fileames; cleaner file setups
