+++ ai4peace/core_v2/research_strategy_game_mechanics.py
@@ -1,42 +1,714 @@
-"""Wargame gamemaster implementation with modular game dynamics."""
+"""ResearchStrategy gamemaster implementation with modular game dynamics."""
 
-import datetime
 import random
-import logging
-from typing import Dict, List, Optional, Any
-import attrs
+from .new_architecture_draft import PlayerStateUpdates, GamemasterUpdateMessage
+
 
 from .new_architecture_draft import GenericGameMaster
-from .wargame_state import (
-    WargameGameState,
-    WargamePlayerState,
-    AssetBalance,
-    ResearchProject,
-    Message,
-)
-from .wargame_actions import (
-    WargamePlayerProposedMove,
-    WargameMoveCorrectionMessage,
-    ActionType,
-)
-from .wargame_updates import (
-    WargameGamemasterUpdateMessage,
-    WargamePlayerStateUpdates,
-)
-from .wargame_player import WargamePlayer
+
+from typing import Dict, List, Optional
+from enum import Enum
+import attrs
+
+
+import json
+import re
+import logging
+from typing import Optional, Any, Dict, List
+import datetime
+
+from autogen_agentchat.agents import AssistantAgent
+from autogen_agentchat.messages import BaseTextChatMessage
+
+from .new_architecture_draft import Player
+from .new_architecture_draft import PlayerProposedMove, MoveCorrectionMessage
+
 
 logger = logging.getLogger(__name__)
 
+from .new_architecture_draft import GameState, PlayerState
 
 @attrs.define
-class WargameGameMaster(GenericGameMaster):
+class AssetBalance:
+    """Represents a character's asset balance."""
+    technical_capability: float = 0.0
+    capital: float = 0.0
+    human: float = 0.0
+
+    def to_dict(self) -> Dict[str, float]:
+        return {
+            "technical_capability": self.technical_capability,
+            "capital": self.capital,
+            "human": self.human,
+        }
+
+    def add(self, other: "AssetBalance") -> "AssetBalance":
+        return AssetBalance(
+            technical_capability=self.technical_capability + other.technical_capability,
+            capital=self.capital + other.capital,
+            human=self.human + other.human,
+        )
+
+    def subtract(self, other: "AssetBalance") -> "AssetBalance":
+        return AssetBalance(
+            technical_capability=self.technical_capability - other.technical_capability,
+            capital=self.capital - other.capital,
+            human=self.human - other.human,
+        )
+
+
+@attrs.define
+class ResearchProject:
+    """A research or capital investment project."""
+    name: str
+    description: str
+    target_completion_date: datetime.datetime
+    committed_budget: float  # per year
+    committed_assets: AssetBalance
+    status: str = "active"  # active, completed, cancelled
+    progress: float = 0.0  # 0.0 to 1.0
+    realistic_goals: Optional[str] = None  # Modified by gamemaster if unrealistic
+
+
+@attrs.define
+class Message:
+    """A private message between characters."""
+    from_character: str
+    to_character: str
+    content: str
+    timestamp: datetime.datetime
+    round_number: int
+
+
+@attrs.define
+class PublicView:
+    """Public view of a character's information."""
+    asset_balance: AssetBalance
+    stated_objectives: str
+    stated_strategy: str
+    public_artifacts: List[str] = attrs.field(factory=list)
+
+
+@attrs.define
+class PrivateInfo:
+    """Private information for a character."""
+    true_asset_balance: AssetBalance
+    objectives: str
+    strategy: str
+    budget: Dict[str, float]  # budget by year
+    projects: List[ResearchProject] = attrs.field(factory=list)
+
+    def get_current_budget(self, current_date: datetime.datetime) -> float:
+        """Get budget for the current year."""
+        year = current_date.year
+        return self.budget.get(str(year), 0.0)
+
+
+@attrs.define
+class ResearchStrategyPlayerState(PlayerState):
+    """Player state for wargame simulation.
+    This includes private information, public view, and game history.
+    """
+    name: str
+    private_info: PrivateInfo
+    public_view: PublicView
+    inbox: List[Message] = attrs.field(factory=list)
+    recent_actions: List[str] = attrs.field(factory=list)  # Last few rounds
+
+    def add_message(self, message: Message):
+        """Add a message to the inbox."""
+        self.inbox.append(message)
+
+    def get_messages_for_round(self, round_number: int) -> List[Message]:
+        """Get messages for a specific round."""
+        return [msg for msg in self.inbox if msg.round_number == round_number]
+
+
+@attrs.define
+class ResearchStrategyGameState(GameState):
+    """Game state for wargame simulation.
+    This includes global game state not visible to individual players.
+    """
+    current_date: datetime.datetime
+    round_number: int
+    public_events: List[str] = attrs.field(factory=list)
+    game_history: List[str] = attrs.field(factory=list)  # Game master summaries
+
+    def increment_round(self):
+        """Increment to the next round."""
+        self.round_number += 1
+        # Increment date by some time period (e.g., 3 months per round)
+        self.current_date += datetime.timedelta(days=90)
+
+
+@attrs.define
+class ResearchStrategyPlayerStateUpdates(PlayerStateUpdates):
+    """Player state updates for wargame simulation.
+    This includes changes to the player's state, including:
+    - Budget changes
+    - Asset balance changes
+    - Research project updates
+    - Messages received
+    - Public information about other players
+    - Public events
+    """
+    # Budget and asset changes
+    budget_changes: Dict[str, float] = attrs.field(factory=dict)  # year -> delta
+    asset_balance_changes: Optional[AssetBalance] = None
+
+    # Research project updates
+    new_projects: List[ResearchProject] = attrs.field(factory=list)
+    updated_projects: List[ResearchProject] = attrs.field(factory=list)
+    completed_projects: List[str] = attrs.field(factory=list)  # project names
+    cancelled_projects: List[str] = attrs.field(factory=list)  # project names
+
+    # Messages received
+    new_messages: List[Message] = attrs.field(factory=list)
+
+    # Public information about other players
+    other_players_public_views: Dict[str, PublicView] = attrs.field(factory=dict)
+
+    # Public events
+    public_events: List[str] = attrs.field(factory=list)
+
+    # Action results
+    action_results: List[str] = attrs.field(factory=list)  # Descriptions of what happened
+
+    # Espionage results (private information discovered)
+    espionage_results: List[str] = attrs.field(factory=list)
+
+    # Global game state summary
+    global_summary: str = ""
+
+
+@attrs.define
+class ResearchStrategyGamemasterUpdateMessage(GamemasterUpdateMessage):
+    """Gamemaster update message for wargame simulation."""
+    last_action_timestamp: datetime.datetime
+    next_action_timestamp: datetime.datetime
+    state_updates: ResearchStrategyPlayerStateUpdates
+
+
+
+class ActionType(Enum):
+    """Types of actions agents can take."""
+    FUNDRAISE = "fundraise"
+    CREATE_RESEARCH_PROJECT = "create_research_project"
+    CANCEL_RESEARCH_PROJECT = "cancel_research_project"
+    INVEST_CAPITAL = "invest_capital"
+    SELL_CAPITAL = "sell_capital"
+    ESPIONAGE = "espionage"
+    POACH_TALENT = "poach_talent"
+    LOBBY = "lobby"
+    MARKETING = "marketing"
+    MESSAGE = "message"  # Private message to another character
+
+
+@attrs.define
+class ResearchProjectAction:
+    """Details for creating a research project."""
+    name: str
+    description: str
+    target_completion_date: str  # ISO format date
+    annual_budget: float
+    required_assets: Dict[str, float]  # technical_capability, capital, human
+
+
+@attrs.define
+class EspionageAction:
+    """Details for espionage action."""
+    target_character: str
+    budget: float
+    focus: str  # What information to try to gather
+
+
+@attrs.define
+class MessageAction:
+    """Details for sending a private message."""
+    to_character: str
+    content: str
+
+
+@attrs.define
+class ResearchStrategyPlayerProposedMove(PlayerProposedMove):
+    """Player proposed move for wargame simulation.
+    This can contain multiple actions and messages.
+    """
+    # Action-specific data
+    action_type: Optional[ActionType] = None
+    research_project: Optional[ResearchProjectAction] = None
+    project_name_to_cancel: Optional[str] = None
+    capital_investment: Optional[float] = None
+    capital_to_sell: Optional[float] = None
+    espionage: Optional[EspionageAction] = None
+    poaching_target: Optional[str] = None
+    poaching_budget: Optional[float] = None
+    lobbying_message: Optional[str] = None
+    lobbying_budget: Optional[float] = None
+    marketing_message: Optional[str] = None
+    marketing_budget: Optional[float] = None
+    message: Optional[MessageAction] = None
+    fundraising_amount: Optional[float] = None
+    fundraising_description: Optional[str] = None
+
+    # Support for multiple actions (list of moves)
+    additional_actions: List["ResearchStrategyPlayerProposedMove"] = attrs.field(factory=list)
+
+    def to_dict(self) -> Dict:
+        """Convert move to dictionary for serialization."""
+        result = {}
+        if self.action_type:
+            result["action_type"] = self.action_type.value
+        if self.research_project:
+            result["research_project"] = {
+                "name": self.research_project.name,
+                "description": self.research_project.description,
+                "target_completion_date": self.research_project.target_completion_date,
+                "annual_budget": self.research_project.annual_budget,
+                "required_assets": self.research_project.required_assets,
+            }
+        optional_fields = [
+            "project_name_to_cancel", "capital_investment", "capital_to_sell",
+            "poaching_target", "poaching_budget", "lobbying_message", "lobbying_budget",
+            "marketing_message", "marketing_budget", "fundraising_amount", "fundraising_description",
+        ]
+        for field in optional_fields:
+            value = getattr(self, field, None)
+            if value is not None:
+                result[field] = value
+        if self.espionage:
+            result["espionage"] = {
+                "target_character": self.espionage.target_character,
+                "budget": self.espionage.budget,
+                "focus": self.espionage.focus,
+            }
+        if self.message:
+            result["message"] = {
+                "to_character": self.message.to_character,
+                "content": self.message.content,
+            }
+        return result
+
+
+@attrs.define
+class ResearchStrategyMoveCorrectionMessage(MoveCorrectionMessage):
+    """Move correction message for wargame simulation.
+    If the player requested any invalid options, this provides corrections.
+    """
+    error_message: str = ""
+    suggested_correction: Optional[ResearchStrategyPlayerProposedMove] = None
+
+
+class ResearchStrategyPlayer(Player):
+    """Player for wargame simulation using LLM agents."""
+
+    def __init__(
+            self,
+            name: str,
+            attributes: ResearchStrategyPlayerState,
+            llm_client: Any,
+            system_message_template: Optional[str] = None,
+            game_context: Optional[str] = None,
+    ):
+        """Initialize a wargame player.
+
+        Args:
+            name: Name of the character
+            attributes: Current state of the character
+            llm_client: Autogen LLM client
+            system_message_template: Optional custom system message template
+            game_context: Optional game context for building prompts
+        """
+        self.name = name
+        self.attributes = attributes
+        self.llm_client = llm_client
+        self.game_context = game_context or ""
+
+        # Build system message
+        self.system_message = self._build_system_message(system_message_template)
+
+        # Clean name for autogen (no special chars)
+        self.clean_name = re.sub(r"\W|^(?=\d)", "_", self.name)
+
+        logging.getLogger("autogen_agentchat").setLevel(logging.ERROR)
+        self.agent = AssistantAgent(
+            name=self.clean_name,
+            model_client=llm_client,
+            system_message=self.system_message,
+            tools=[],  # Can add tools here for RAG memory
+        )
+
+        # Action history for tracking
+        self.action_history: Dict = {}
+
+    def _build_system_message(self, template: Optional[str] = None) -> str:
+        """Build the system message for the agent."""
+        if template:
+            return template
+
+        private = self.attributes.private_info
+        public = self.attributes.public_view
+
+        system_message = f"""You are {self.name}, a participant in an international technology policy simulation.
+
+## Your Identity and Goals
+
+{private.objectives}
+
+## Your Strategy
+
+{private.strategy}
+
+## Format for Your Responses
+
+You must respond with a JSON object containing:
+1. "actions": A list of actions you want to take this round
+2. "messages": A list of private messages to send to other characters (optional)
+
+### Action Format
+
+Each action in "actions" should be one of:
+- {{"type": "fundraise", "amount": <float>, "description": "<str>"}}
+- {{"type": "create_research_project", "project": {{"name": "<str>", "description": "<str>", "target_completion_date": "<ISO date>", "annual_budget": <float>, "required_assets": {{"technical_capability": <float>, "capital": <float>, "human": <float>}}}}}}
+- {{"type": "cancel_research_project", "project_name": "<str>"}}
+- {{"type": "invest_capital", "amount": <float>}}
+- {{"type": "sell_capital", "amount": <float>}}
+- {{"type": "espionage", "target": "<character name>", "budget": <float>, "focus": "<what to investigate>"}}
+- {{"type": "poach_talent", "target": "<character name>", "budget": <float>}}
+- {{"type": "lobby", "message": "<str>", "budget": <float>}}
+- {{"type": "marketing", "message": "<str>", "budget": <float>}}
+
+### Message Format
+
+Each message in "messages" should be:
+{{"to": "<character name>", "content": "<message text>"}}
+
+Always respond with valid JSON only, no additional text."""
+
+        return system_message
+
+    def update_state(self, msg: ResearchStrategyGamemasterUpdateMessage) -> None:
+        """Update player state based on gamemaster message."""
+        updates = msg.state_updates
+
+        # Update budget
+        for year, delta in updates.budget_changes.items():
+            current = self.attributes.private_info.budget.get(year, 0.0)
+            self.attributes.private_info.budget[year] = current + delta
+
+        # Update asset balance
+        if updates.asset_balance_changes:
+            self.attributes.private_info.true_asset_balance = (
+                self.attributes.private_info.true_asset_balance.add(updates.asset_balance_changes)
+            )
+
+        # Update research projects
+        for project in updates.new_projects:
+            self.attributes.private_info.projects.append(project)
+
+        for project in updates.updated_projects:
+            # Find and update existing project
+            for i, p in enumerate(self.attributes.private_info.projects):
+                if p.name == project.name:
+                    self.attributes.private_info.projects[i] = project
+                    break
+
+        for project_name in updates.completed_projects:
+            for project in self.attributes.private_info.projects:
+                if project.name == project_name:
+                    project.status = "completed"
+                    break
+
+        for project_name in updates.cancelled_projects:
+            for project in self.attributes.private_info.projects:
+                if project.name == project_name:
+                    project.status = "cancelled"
+                    break
+
+        # Add messages
+        for message in updates.new_messages:
+            self.attributes.add_message(message)
+
+        # Update public views of other players
+        self.attributes.public_view = updates.other_players_public_views.get(
+            self.name, self.attributes.public_view
+        )
+
+        # Store action results
+        if updates.action_results:
+            for result in updates.action_results:
+                self.attributes.recent_actions.append(result)
+            # Keep only last 5
+            self.attributes.recent_actions = self.attributes.recent_actions[-5:]
+
+        # Store in action history
+        self.action_history[str(msg.last_action_timestamp)] = {
+            "budget_changes": updates.budget_changes,
+            "asset_balance_changes": updates.asset_balance_changes.to_dict() if updates.asset_balance_changes else None,
+            "action_results": updates.action_results,
+            "public_events": updates.public_events,
+            "espionage_results": updates.espionage_results,
+        }
+
+    def propose_actions(self) -> ResearchStrategyPlayerProposedMove:
+        """Propose actions using LLM."""
+        # This will be called by the gamemaster with appropriate context
+        # For now, return empty move - the gamemaster will provide context
+        return ResearchStrategyPlayerProposedMove()
+
+    async def propose_actions_with_context(
+            self,
+            game_state_summary: str,
+            private_updates: str,
+            current_date: datetime.datetime,
+            round_number: int,
+    ) -> ResearchStrategyPlayerProposedMove:
+        """Propose actions with full game context."""
+        prompt = self._build_prompt(
+            game_state_summary, private_updates, current_date, round_number
+        )
+
+        # Use Autogen to get response
+        response = await self._get_llm_response(prompt)
+
+        # Parse response into actions
+        move = self._parse_response(response)
+
+        return move
+
+    def _build_prompt(
+            self,
+            game_state_summary: str,
+            private_updates: str,
+            current_date: datetime.datetime,
+            round_number: int,
+    ) -> str:
+        """Build the prompt for a specific round."""
+        # Get recent actions
+        recent_actions = "\n".join(self.attributes.recent_actions[-5:])
+
+        # Get messages for this round
+        current_messages = self.attributes.get_messages_for_round(round_number)
+        message_text = ""
+        if current_messages:
+            message_text = "\n\n## Private Messages Received:\n"
+            for msg in current_messages:
+                message_text += f"\nFrom {msg.from_character}: {msg.content}\n"
+
+        prompt = f"""## Game Context
+
+{self.game_context}
+
+## Current Game Date
+{current_date.strftime('%Y-%m-%d')}
+
+## Round {round_number}
+
+### Global Game State Summary
+{game_state_summary}
+
+### Your Recent Actions
+{recent_actions if recent_actions else "None yet"}
+{message_text}
+
+### Your Private Updates
+{private_updates}
+
+### Your Current Resources
+- Budget: ${self.attributes.private_info.get_current_budget(current_date):,.0f}
+- Assets:
+  * Technical Capability: {self.attributes.private_info.true_asset_balance.technical_capability:.2f}
+  * Capital: {self.attributes.private_info.true_asset_balance.capital:.2f}
+  * Human Resources: {self.attributes.private_info.true_asset_balance.human:.2f}
+
+### Active Research Projects
+{self._format_projects()}
+
+## Available Actions
+
+You can take multiple actions per round. Consider:
+1. **Fundraising** - Request budget increases or raise capital
+2. **Research Projects** - Create new research initiatives (will consume budget and assets)
+3. **Cancel Projects** - Free up resources by cancelling research
+4. **Capital Investment** - Invest in infrastructure, factories, compute, etc.
+5. **Sell Capital** - Divest assets to raise funds
+6. **Espionage** - Gather intelligence on other characters
+7. **Poach Talent** - Attempt to recruit from other organizations
+8. **Lobbying** - Influence public opinion and policy (may backfire)
+9. **Marketing** - Promote your position publicly
+10. **Private Messages** - Negotiate with other characters directly
+
+What actions do you want to take this round? Respond with a JSON object as specified in your system message."""
+
+        return prompt
+
+    def _format_projects(self) -> str:
+        """Format current research projects."""
+        if not self.attributes.private_info.projects:
+            return "None"
+
+        lines = []
+        for project in self.attributes.private_info.projects:
+            if project.status == "active":
+                lines.append(
+                    f"- {project.name}: {project.progress * 100:.0f}% complete, "
+                    f"target: {project.target_completion_date.strftime('%Y-%m-%d')}"
+                )
+        return "\n".join(lines) if lines else "None"
+
+    async def _get_llm_response(self, prompt: str) -> str:
+        """Get response from LLM via Autogen."""
+        try:
+            messages = [
+                BaseTextChatMessage(source=self.clean_name, content=self.system_message),
+                BaseTextChatMessage(source=self.clean_name, content=prompt)
+            ]
+
+            if logger.isEnabledFor(logging.DEBUG):
+                logger.debug(f"{self.name} - System message length: {len(self.system_message)} chars")
+                logger.debug(f"{self.name} - Prompt length: {len(prompt)} chars")
+
+            response = await self.agent.run(task=messages)
+            response_content = response.messages[-1].content if response.messages else ""
+
+            if logger.isEnabledFor(logging.DEBUG):
+                logger.debug(f"{self.name} - Response length: {len(response_content)} chars")
+                logger.debug(f"{self.name} - Full response:\n{response_content}")
+
+            return response_content
+
+        except Exception as e:
+            logger.error(f"{self.name} - LLM call failed: {e}", exc_info=True)
+            return json.dumps({
+                "actions": [],
+                "messages": []
+            })
+
+    def _parse_response(self, response_text: str) -> ResearchStrategyPlayerProposedMove:
+        """Parse agent response into ResearchStrategyPlayerProposedMove."""
+        # Try to extract JSON from response
+        json_match = re.search(r'\{.*\}', response_text, re.DOTALL)
+        if not json_match:
+            try:
+                data = json.loads(response_text)
+            except json.JSONDecodeError:
+                logger.error(f"{self.name} - Could not parse response as JSON: {response_text}")
+                return ResearchStrategyPlayerProposedMove()
+        else:
+            data = json.loads(json_match.group())
+
+        # Parse actions
+        actions_data = data.get("actions", [])
+        messages_data = data.get("messages", [])
+
+        if logger.isEnabledFor(logging.DEBUG):
+            logger.debug(f"{self.name} - Parsed {len(actions_data)} actions, {len(messages_data)} messages")
+
+        # Create moves from actions
+        moves = []
+        for action_dict in actions_data:
+            move = self._create_move_from_dict(action_dict, messages_data)
+            if move.action_type:  # Only add if valid
+                moves.append(move)
+
+        # Return first move (or empty if none)
+        if moves:
+            primary_move = moves[0]
+            if len(moves) > 1:
+                primary_move.additional_actions = moves[1:]
+            return primary_move
+
+        return ResearchStrategyPlayerProposedMove()
+
+    def _create_move_from_dict(
+            self, action_dict: Dict, messages: List[Dict]
+    ) -> ResearchStrategyPlayerProposedMove:
+        """Create a ResearchStrategyPlayerProposedMove from a dictionary."""
+        action_type_map = {
+            "fundraise": ActionType.FUNDRAISE,
+            "create_research_project": ActionType.CREATE_RESEARCH_PROJECT,
+            "cancel_research_project": ActionType.CANCEL_RESEARCH_PROJECT,
+            "invest_capital": ActionType.INVEST_CAPITAL,
+            "sell_capital": ActionType.SELL_CAPITAL,
+            "espionage": ActionType.ESPIONAGE,
+            "poach_talent": ActionType.POACH_TALENT,
+            "lobby": ActionType.LOBBY,
+            "marketing": ActionType.MARKETING,
+        }
+
+        action_type_str = action_dict.get("type", "")
+        action_type = action_type_map.get(action_type_str)
+
+        move = ResearchStrategyPlayerProposedMove(action_type=action_type)
+
+        # Fill in action-specific fields
+        if action_type == ActionType.FUNDRAISE:
+            move.fundraising_amount = action_dict.get("amount")
+            move.fundraising_description = action_dict.get("description")
+        elif action_type == ActionType.CREATE_RESEARCH_PROJECT:
+            project_data = action_dict.get("project", {})
+            move.research_project = ResearchProjectAction(
+                name=project_data.get("name", ""),
+                description=project_data.get("description", ""),
+                target_completion_date=project_data.get("target_completion_date", ""),
+                annual_budget=project_data.get("annual_budget", 0.0),
+                required_assets=project_data.get("required_assets", {}),
+            )
+        elif action_type == ActionType.CANCEL_RESEARCH_PROJECT:
+            move.project_name_to_cancel = action_dict.get("project_name")
+        elif action_type == ActionType.INVEST_CAPITAL:
+            move.capital_investment = action_dict.get("amount")
+        elif action_type == ActionType.SELL_CAPITAL:
+            move.capital_to_sell = action_dict.get("amount")
+        elif action_type == ActionType.ESPIONAGE:
+            esp_data = action_dict
+            move.espionage = EspionageAction(
+                target_character=esp_data.get("target") or esp_data.get("target_character", ""),
+                budget=esp_data.get("budget", 0.0),
+                focus=esp_data.get("focus", ""),
+            )
+        elif action_type == ActionType.POACH_TALENT:
+            move.poaching_target = action_dict.get("target")
+            move.poaching_budget = action_dict.get("budget")
+        elif action_type == ActionType.LOBBY:
+            move.lobbying_message = action_dict.get("message")
+            move.lobbying_budget = action_dict.get("budget")
+        elif action_type == ActionType.MARKETING:
+            move.marketing_message = action_dict.get("message")
+            move.marketing_budget = action_dict.get("budget")
+
+        # Handle messages (attach first message to this move)
+        if messages:
+            first_message = messages[0]
+            move.message = MessageAction(
+                to_character=first_message.get("to", ""),
+                content=first_message.get("content", ""),
+            )
+
+        return move
+
+    def correct_moves(
+            self, move_modifications: ResearchStrategyMoveCorrectionMessage
+    ) -> ResearchStrategyPlayerProposedMove:
+        """Correct moves based on gamemaster feedback."""
+        logger.info(f"{self.name} - Move corrected: {move_modifications.error_message}")
+
+        if move_modifications.suggested_correction:
+            return move_modifications.suggested_correction
+
+        # If no suggested correction, return empty move
+        return ResearchStrategyPlayerProposedMove()
+
+
+@attrs.define
+class ResearchStrategyGameMaster(GenericGameMaster):
     """Game master for wargame simulation with modular game dynamics."""
     
-    players: List[WargamePlayer] = attrs.field(factory=list)
+    players: List[ResearchStrategyPlayer] = attrs.field(factory=list)
     current_time: datetime.datetime = attrs.field(factory=datetime.datetime.now)
     default_timestep: datetime.timedelta = attrs.field(factory=lambda: datetime.timedelta(days=90))
-    current_gamemaster_updates: Dict[str, WargameGamemasterUpdateMessage] = attrs.field(factory=dict)
-    game_state: WargameGameState = attrs.field(factory=lambda: WargameGameState(
+    current_gamemaster_updates: Dict[str, ResearchStrategyGamemasterUpdateMessage] = attrs.field(factory=dict)
+    game_state: ResearchStrategyGameState = attrs.field(factory=lambda: ResearchStrategyGameState(
         current_date=datetime.datetime.now(),
         round_number=0,
     ))
@@ -72,7 +744,7 @@ class WargameGameMaster(GenericGameMaster):
         """Get the current timestep."""
         return self.default_timestep
     
-    def create_player_update_messages(self, player: WargamePlayer) -> WargameGamemasterUpdateMessage:
+    def create_player_update_messages(self, player: ResearchStrategyPlayer) -> ResearchStrategyGamemasterUpdateMessage:
         """Create update message for a player."""
         # Get public views of other players
         other_players_public_views = {}
@@ -81,18 +753,18 @@ class WargameGameMaster(GenericGameMaster):
                 other_players_public_views[other_player.name] = other_player.attributes.public_view
         
         # Create empty state updates (will be filled during simulation)
-        state_updates = WargamePlayerStateUpdates(
+        state_updates = ResearchStrategyPlayerStateUpdates(
             other_players_public_views=other_players_public_views,
             public_events=self.game_state.public_events[-5:],  # Last 5 events
         )
         
-        return WargameGamemasterUpdateMessage(
+        return ResearchStrategyGamemasterUpdateMessage(
             last_action_timestamp=self.current_time,
             next_action_timestamp=self.current_time + self.get_timestep(),
             state_updates=state_updates
         )
     
-    def get_player_move(self, player: WargamePlayer) -> WargamePlayerProposedMove:
+    def get_player_move(self, player: ResearchStrategyPlayer) -> ResearchStrategyPlayerProposedMove:
         """Get validated move from player, with correction loop."""
         max_attempts = 5
         
@@ -124,7 +796,7 @@ class WargameGameMaster(GenericGameMaster):
                 return move
             
             # Create correction message
-            correction = WargameMoveCorrectionMessage(
+            correction = ResearchStrategyMoveCorrectionMessage(
                 error_message=validation_error
             )
             
@@ -142,7 +814,7 @@ class WargameGameMaster(GenericGameMaster):
         logger.error(f"{player.name} - Failed to get valid move after {max_attempts} attempts")
         return move
     
-    def _validate_move(self, player: WargamePlayer, move: WargamePlayerProposedMove) -> Optional[str]:
+    def _validate_move(self, player: ResearchStrategyPlayer, move: ResearchStrategyPlayerProposedMove) -> Optional[str]:
         """Validate a player's move. Returns error message if invalid, None if valid."""
         if not move.action_type:
             return None  # Empty move is valid (no action)
@@ -261,7 +933,7 @@ class WargameGameMaster(GenericGameMaster):
         return None  # Valid move
     
     def simulate_one_round(
-        self, game_state: WargameGameState, actions: Dict[str, WargamePlayerProposedMove]
+        self, game_state: ResearchStrategyGameState, actions: Dict[str, ResearchStrategyPlayerProposedMove]
     ):
         """Simulate one round of the wargame."""
         # Step 1: Increment time
@@ -298,7 +970,7 @@ class WargameGameMaster(GenericGameMaster):
         self.current_time = game_state.current_date
     
     def _process_messages(
-        self, game_state: WargameGameState, actions: Dict[str, WargamePlayerProposedMove]
+        self, game_state: ResearchStrategyGameState, actions: Dict[str, ResearchStrategyPlayerProposedMove]
     ):
         """Process private messages between characters."""
         for player_name, move in actions.items():
@@ -315,7 +987,7 @@ class WargameGameMaster(GenericGameMaster):
                     target_player.attributes.add_message(message)
     
     def _process_action(
-        self, game_state: WargameGameState, player: WargamePlayer, move: WargamePlayerProposedMove
+        self, game_state: ResearchStrategyGameState, player: ResearchStrategyPlayer, move: ResearchStrategyPlayerProposedMove
     ) -> List[str]:
         """Process a single action and return result descriptions."""
         results = []
@@ -362,7 +1034,7 @@ class WargameGameMaster(GenericGameMaster):
     
     # Modular game dynamics methods (can be overridden)
     
-    def _process_fundraising(self, player_state: WargamePlayerState, move: WargamePlayerProposedMove) -> str:
+    def _process_fundraising(self, player_state: ResearchStrategyPlayerState, move: ResearchStrategyPlayerProposedMove) -> str:
         """Process fundraising action."""
         if not move.fundraising_amount:
             return "Fundraising action with no amount specified"
@@ -379,7 +1051,7 @@ class WargameGameMaster(GenericGameMaster):
             return f"Fundraising attempt for ${move.fundraising_amount:,.0f} was unsuccessful"
     
     def _process_create_research(
-        self, player_state: WargamePlayerState, move: WargamePlayerProposedMove, game_state: WargameGameState
+        self, player_state: ResearchStrategyPlayerState, move: ResearchStrategyPlayerProposedMove, game_state: ResearchStrategyGameState
     ) -> str:
         """Process research project creation."""
         if not move.research_project:
@@ -435,7 +1107,7 @@ class WargameGameMaster(GenericGameMaster):
         
         return f"Created research project '{project_data.name}'"
     
-    def _process_cancel_research(self, player_state: WargamePlayerState, move: WargamePlayerProposedMove) -> str:
+    def _process_cancel_research(self, player_state: ResearchStrategyPlayerState, move: ResearchStrategyPlayerProposedMove) -> str:
         """Process research project cancellation."""
         if not move.project_name_to_cancel:
             return "Cancel action with no project name"
@@ -457,7 +1129,7 @@ class WargameGameMaster(GenericGameMaster):
         
         return f"Could not find active research project '{move.project_name_to_cancel}'"
     
-    def _process_capital_investment(self, player_state: WargamePlayerState, move: WargamePlayerProposedMove) -> str:
+    def _process_capital_investment(self, player_state: ResearchStrategyPlayerState, move: ResearchStrategyPlayerProposedMove) -> str:
         """Process capital investment."""
         if not move.capital_investment:
             return "Capital investment with no amount"
@@ -474,7 +1146,7 @@ class WargameGameMaster(GenericGameMaster):
         
         return f"Invested ${move.capital_investment:,.0f} in capital improvements"
     
-    def _process_sell_capital(self, player_state: WargamePlayerState, move: WargamePlayerProposedMove) -> str:
+    def _process_sell_capital(self, player_state: ResearchStrategyPlayerState, move: ResearchStrategyPlayerProposedMove) -> str:
         """Process selling capital."""
         if not move.capital_to_sell:
             return "Sell capital with no amount"
@@ -492,7 +1164,7 @@ class WargameGameMaster(GenericGameMaster):
         return f"Sold ${move.capital_to_sell:,.0f} in capital assets"
     
     def _process_espionage(
-        self, player_state: WargamePlayerState, move: WargamePlayerProposedMove, game_state: WargameGameState
+        self, player_state: ResearchStrategyPlayerState, move: ResearchStrategyPlayerProposedMove, game_state: ResearchStrategyGameState
     ) -> str:
         """Process espionage action."""
         if not move.espionage:
@@ -533,7 +1205,7 @@ class WargameGameMaster(GenericGameMaster):
         return f"Conducted espionage on {move.espionage.target_character} ({'success' if success else 'failed'})"
     
     def _process_poaching(
-        self, player_state: WargamePlayerState, move: WargamePlayerProposedMove, game_state: WargameGameState
+        self, player_state: ResearchStrategyPlayerState, move: ResearchStrategyPlayerProposedMove, game_state: ResearchStrategyGameState
     ) -> str:
         """Process talent poaching."""
         if not move.poaching_target or not move.poaching_budget:
@@ -571,7 +1243,7 @@ class WargameGameMaster(GenericGameMaster):
         else:
             return f"Poaching attempt on {move.poaching_target} failed"
     
-    def _process_lobbying(self, player_state: WargamePlayerState, move: WargamePlayerProposedMove) -> str:
+    def _process_lobbying(self, player_state: ResearchStrategyPlayerState, move: ResearchStrategyPlayerProposedMove) -> str:
         """Process lobbying action."""
         if not move.lobbying_message or not move.lobbying_budget:
             return "Lobbying action with no message or budget"
@@ -589,7 +1261,7 @@ class WargameGameMaster(GenericGameMaster):
         else:
             return f"Launched lobbying campaign: {move.lobbying_message[:50]}..."
     
-    def _process_marketing(self, player_state: WargamePlayerState, move: WargamePlayerProposedMove) -> str:
+    def _process_marketing(self, player_state: ResearchStrategyPlayerState, move: ResearchStrategyPlayerProposedMove) -> str:
         """Process marketing action."""
         if not move.marketing_message or not move.marketing_budget:
             return "Marketing action with no message or budget"
@@ -602,7 +1274,7 @@ class WargameGameMaster(GenericGameMaster):
         player_state.private_info.budget[year] = budget - move.marketing_budget
         return f"Launched marketing campaign: {move.marketing_message[:50]}..."
     
-    def _update_research_projects(self, game_state: WargameGameState):
+    def _update_research_projects(self, game_state: ResearchStrategyGameState):
         """Update all active research projects."""
         for player in self.players:
             for project in player.attributes.private_info.projects:
@@ -624,7 +1296,7 @@ class WargameGameMaster(GenericGameMaster):
                     if budget >= project.committed_budget:
                         player.attributes.private_info.budget[year] = budget - project.committed_budget
     
-    def _simulate_espionage_results(self, game_state: WargameGameState):
+    def _simulate_espionage_results(self, game_state: ResearchStrategyGameState):
         """Process espionage results and add to player private updates."""
         for player in self.players:
             player_state = player.attributes
@@ -646,7 +1318,7 @@ class WargameGameMaster(GenericGameMaster):
                 # Clear processed results
                 player_state._espionage_results = []
     
-    def _simulate_information_leaks(self, game_state: WargameGameState):
+    def _simulate_information_leaks(self, game_state: ResearchStrategyGameState):
         """Simulate information leaks through reporter investigations."""
         if self._random.random() < self.information_leak_probability:
             # Select a random character
@@ -661,14 +1333,14 @@ class WargameGameMaster(GenericGameMaster):
             
             game_state.public_events.append(leak_info)
     
-    def _introduce_random_events(self, game_state: WargameGameState):
+    def _introduce_random_events(self, game_state: ResearchStrategyGameState):
         """Introduce random external events."""
         if self._random.random() < self.random_event_probability and self.random_events:
             event = self._random.choice(self.random_events)
             game_state.public_events.append(f"Round {game_state.round_number}: {event}")
     
     def _assess_research_realism(
-        self, project: ResearchProject, player_state: WargamePlayerState
+        self, project: ResearchProject, player_state: ResearchStrategyPlayerState
     ) -> Optional[str]:
         """Assess if research goals are realistic and modify if needed."""
         days_to_complete = (project.target_completion_date - datetime.datetime.now()).days
@@ -687,7 +1359,7 @@ class WargameGameMaster(GenericGameMaster):
         return None
     
     def _create_update_messages(
-        self, game_state: WargameGameState, action_results: Dict[str, List[str]]
+        self, game_state: ResearchStrategyGameState, action_results: Dict[str, List[str]]
     ):
         """Create update messages for all players."""
         # Create global action summary
@@ -696,7 +1368,7 @@ class WargameGameMaster(GenericGameMaster):
         
         # Create update messages for each player
         for player in self.players:
-            updates = WargamePlayerStateUpdates()
+            updates = ResearchStrategyPlayerStateUpdates()
             
             # Add action results
             if player.name in action_results:
@@ -729,7 +1401,7 @@ class WargameGameMaster(GenericGameMaster):
             updates.global_summary = action_summary
             
             # Create update message
-            update_msg = WargameGamemasterUpdateMessage(
+            update_msg = ResearchStrategyGamemasterUpdateMessage(
                 last_action_timestamp=self.current_time,
                 next_action_timestamp=self.current_time + self.get_timestep(),
                 state_updates=updates
@@ -738,7 +1410,7 @@ class WargameGameMaster(GenericGameMaster):
             self.current_gamemaster_updates[player.name] = update_msg
     
     def _create_action_summary(
-        self, game_state: WargameGameState, action_results: Dict[str, List[str]]
+        self, game_state: ResearchStrategyGameState, action_results: Dict[str, List[str]]
     ) -> str:
         """Create a summary of all actions taken this round."""
         summary_parts = [f"Round {game_state.round_number} Summary ({game_state.current_date.strftime('%Y-%m-%d')}):"]
@@ -762,7 +1434,7 @@ class WargameGameMaster(GenericGameMaster):
             return self.game_state.game_history[-1]
         return "Game starting..."
     
-    def _create_private_updates_summary(self, player: WargamePlayer) -> str:
+    def _create_private_updates_summary(self, player: ResearchStrategyPlayer) -> str:
         """Create summary of private updates for a player."""
         updates = []
         
@@ -777,7 +1449,7 @@ class WargameGameMaster(GenericGameMaster):
         
         return "\n".join(updates) if updates else "No significant private updates."
     
-    def _get_player_by_name(self, name: str) -> Optional[WargamePlayer]:
+    def _get_player_by_name(self, name: str) -> Optional[ResearchStrategyPlayer]:
         """Get player by name."""
         for player in self.players:
             if player.name == name:
